+++ tutor/templates/apps/openedx/settings/cms/__init__.py
@@ -1,0 +1,1 @@
+{% include "apps/openedx/settings/partials/pre_common_all.py" %}