+++ tutor/templates/apps/openedx/settings/cms/production.py
@@ -9,6 +9,4 @@ ALLOWED_HOSTS = [
     "cms",
 ]
 
-MIDDLEWARE.insert(0, "whitenoise.middleware.WhiteNoiseMiddleware")
-
 {{ patch("openedx-cms-production-settings") }}