+++ tutor/templates/apps/openedx/settings/partials/pre_common_all.py
@@ -1,0 +1,10 @@
+# Silence overly verbose warnings
+import logging
+import warnings
+from django.utils.deprecation import RemovedInDjango30Warning, RemovedInDjango31Warning
+from rest_framework import RemovedInDRF310Warning, RemovedInDRF311Warning
+warnings.simplefilter('ignore', RemovedInDjango30Warning)
+warnings.simplefilter('ignore', RemovedInDjango31Warning)
+warnings.simplefilter('ignore', RemovedInDRF310Warning)
+warnings.simplefilter('ignore', RemovedInDRF311Warning)
+warnings.simplefilter('ignore', DeprecationWarning)